significantly speed up deletion when deleting a large range of cells
