+++ src/gridGL/interaction/CellInput.tsx
@@ -6,7 +6,7 @@ import { PixiApp } from '../pixiApp/PixiApp';
 import { localFiles } from '../../grid/sheet/localFiles';
 import { SheetController } from '../../grid/controller/sheetController';
 import { updateCellAndDCells } from '../../grid/actions/updateCellAndDCells';
-import { DeleteCells } from '../../grid/sheet/Cells/DeleteCells';
+import { DeleteCells } from '../../grid/actions/DeleteCells';
 import { EditorInteractionState } from '../../atoms/editorInteractionStateAtom';
 
 interface CellInputProps {
